+++ src/main.h
@@ -18,6 +18,7 @@
 #ifndef __MAIN_H__
 #define __MAIN_H__
 
+#include <stddef.h>
 #include <wonderful.h>
 #include <ws.h>
 
@@ -31,7 +32,7 @@ void tile_copy_2bpp_to_4bpp(uint8_t __wf_iram* dest, const uint8_t __wf_rom* src
 
 // 8-high variable width font system
 int vwf8_get_string_width(const char __wf_rom* s);
-int vwf8_draw_char(uint16_t __wf_iram* tile, uint8_t chr, int x);
-int vwf8_draw_string(uint16_t __wf_iram* tile, const char __wf_rom* s, int x);
+int vwf8_draw_char(uint8_t __wf_iram* tile, uint8_t chr, int x);
+int vwf8_draw_string(uint8_t __wf_iram* tile, const char __wf_rom* s, int x);
 
 #endif
